Move state.json related code into subpackage state
